lxd: Move certificate cache update functions to new file.
